+++ R/globals.R
@@ -1,4 +1,4 @@
 
 
 # Global bindings for package checks
-#utils::globalVariables(c("foo", "bar"))
+utils::globalVariables(c("binwidth"))